Add School Profile page and new routing mechanism
This commit adds a new routing mechanism that does not require routes to
be manually added for each page; instead, it attempts to serve all uris
given to it and redirects to home if they are not available.  As a test
of this routing mechanism, this commit also adds the beginnings of a
school profile page.